+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -418,7 +418,7 @@ class CreateDatasourceResponseBody(BaseModel):
     metadata: CreateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYS1R5AXQW2V4N8Y6DZW536V"
+        "01KYSA4S5NSYQPMZZF80C9PWMH"
     )
     r"""The unique identifier of the data source"""