+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsehttp.py
@@ -55,9 +55,9 @@ class EvaluatorResponseHTTP(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-30T08:19:59.802Z"
+    created: Optional[str] = "2026-07-30T10:46:42.051Z"
 
-    updated: Optional[str] = "2026-07-30T08:19:59.802Z"
+    updated: Optional[str] = "2026-07-30T10:46:42.051Z"
 
     updated_by_id: OptionalNullable[str] = UNSET