+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsellm.py
@@ -250,9 +250,9 @@ class EvaluatorResponseLlm(BaseModel):
 
     mode: Mode
 
-    created: Optional[str] = "2026-07-30T08:19:59.802Z"
+    created: Optional[str] = "2026-07-30T10:46:42.051Z"
 
-    updated: Optional[str] = "2026-07-30T08:19:59.802Z"
+    updated: Optional[str] = "2026-07-30T10:46:42.051Z"
 
     updated_by_id: OptionalNullable[str] = UNSET