+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -247,7 +247,7 @@ class ListDatasourcesData(BaseModel):
     metadata: ListDatasourcesMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYS1R5ASZ6YFR5CMXN3QFYPP"
+        "01KYSA4S5JHCT2Z78A2DTJTBBP"
     )
     r"""The unique identifier of the data source"""