+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -169,7 +169,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     metadata: RetrieveDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYS1R5AV6J3GH1S4KFBRP0C2"
+        "01KYSA4S5MREC1C6X7CF53X27K"
     )
     r"""The unique identifier of the data source"""