+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -160,7 +160,7 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-07-30T08:19:57.737Z")
+    updated: Optional[datetime] = parse_datetime("2026-07-30T10:46:39.932Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")