+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -183,7 +183,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     metadata: UpdateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYS1R5AZT8M6CYXR3NGKAKYM"
+        "01KYSA4S5QSVHANWPVN1NHF4CT"
     )
     r"""The unique identifier of the data source"""