+++ packages/orq-rc/RELEASES.md
@@ -9859,3 +9859,13 @@ Based on:
 - [python v4.13.0-rc.19] packages/orq-rc
 ### Releases
 - [PyPI v4.13.0-rc.19] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.19 - packages/orq-rc
+
+## 2026-07-30 10:50:38
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.791.0 (2.924.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.13.0-rc.20] packages/orq-rc
+### Releases
+- [PyPI v4.13.0-rc.20] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.20 - packages/orq-rc